+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed 🐛
 
 * Fixed a bug that would sometimes trigger a "Mismatching panels" panic in debug builds.
+* `Image` and `ImageButton` will no longer stretch to fill a justified layout.
 
 ## 0.7.0 - 2021-01-04